feat: allow overriding actions in snapshot module
We need to surface the defaults in this module, since otherwise we have
to constantly release a new `terraform-aws-lambda` module every time we
want to update the default collection policy. This leads to some
redundancy in where we compile defaults, but over time this module is
a better home for default values, as users will be less likely to be
interfacing with the individual submodules.