fix: remove action variable, allow falling back to default (#12)
We update the default actions list implicitly every time we bump the
version of the `terraform-aws-lambda` module. Rather than keep tabs on
what actions are supported in two separate modules, always fall back to
whatever the underlying lambda module supports.

In practice this change breaks backward compatibility, but there is no
loss of functionality. The `snapshot_include` and `snapshot_exclude`
variables can be used to tailor the list of actions if needed.