fix!: update cloudwatch logs subscription module
This commit bumps the cloudwatch logs subscription module and gates it
conditionally on the `subscribed_log_group_matches` and
`subscribed_log_group_excludes` variables. If neither list is provided,
the module will not be installed.

As a result of these changes, the minimum terraform provider version is
now 1.1.0. This is matches the requirements of the underyling cloudwatch
logs subscription module, and allows the usage of `moved` directive to
retain existing cloudwatch log subscription modules across upgrades.