fix: respect cloudtrail configuration in eventbridge
This commit ensures that we take into account user preferences for
cloudtrail when forwarding events from eventbridge.

Cloudtrail events are available on the default event bus. Previously,
all cloudtrail events would be forwarded to Observe, independently of
the value of `enable_cloudtrail` or
`cloudtrail_exclude_management_event_sources`, which encode user
preferences for Cloudtrail.

We break the existing eventbridge rule in two:
- a rule that matches all events _other_ than Cloudtrail
- a rule that matches all Cloudtrail excluding management events in
  exclusion list. This rule is only created if `enable_cloudtrail` is
  true.