feat: allow opting out of cloudwatch metrics stream (#48)
This commit allows callers to opt out of configuring cloudwatch metrics
stream.

Prior to this commit, users could provide an inclusion list or exclusion
list of namespaces to subscribe to. The only option available to disable
exporting metrics would be to provide a list of all possible namespaces.
This is a very long list, and you'd still end up with a metrics stream.

We now gate metrics stream configuration on a wildcard in the exclusion
list. This makes things less cumbersome than adding a new bool
variable (e.g. `cloudwatch_metrics_enable`), and more explicit than
retrofitting the existing inclusion list to accept a null.