feat: introduce `s3_bucket_name` variable
This commit introduces the ability to override the bucket name used for
our collection bucket.

By default, we create a bucket name from the concatenation of `name`,
`region`, and some random suffix. The fact this is not known at apply
time can be a hinderance in some cases, e.g. trying to tag a bucket with
its name.