feat(config): allow publishing to SNS topic
Allow AWS Config to publish change notifications to a provided SNS
topic.

We can not use `count` on `var.sns_topic_arn` to gate the construction
of the IAM policy in this case. Terraform will complain it cannot
determine the number of instances prior to apply. Instead we always
declare a well formed policy, but only inline it if an SNS topic is
provided.