feat(forwarder): relax filedrop requirement
Allow forwarder to be set up with an HTTP endpoint. This commit
introduces a `uri` field in `destination`, and makes all fields within
the object optional.

This commit deviates from the cloudformation equivalent parameters. In
cloudformation, we have a single `DestinationUri` parameter which
accepts both `s3` and `https` schemes. In Terraform, we prioritize
ergonomics for passing in an `observe_filedrop` object. As a result,
`uri` is reserved solely for the case where we pass in an `https` URL.