fix(forwarder): adjust defaults based on URI (#146)
Memory usage will vary according to whether we are using filedrop or
sending to HTTP. When sending to HTTP, memory and CPU usage will scale
according to the size of input, whereas in Filedrop the object copy is
essentially scale invariant. We should adjust our defaults accordingly
rather than setting them in variables.