fix(metricstream): set recommended filter by omission
Previously we were collecting all metrics by omission. This commit
instead loads a recommended set of filters from a file on disk. The file
is kept in sync with the filter we provide for cloudformation.

We could have loaded the filter dynamically directly from the hosted
URL, but that would be at odds with the idempotent nature of terraform
modules. Since we already have to update the binary versions of lambda
functions, we can piggyback on that process to mirror the file.