fix(forwarder): relax IAM role naming
The IAM role of the forwarder only needs to be known when forwarding
data to Filedrop. For cases where we are not pointing at an S3 data
access point, we can fallback to provisioning a role with `name_prefix`
rather than `name`. Doing so avoids the possibility of name clashes when
installing a forwarder in different regions.