refactor: replace `samversion` with `sam_asset`
The previous module was overly restrictive, by assuming that a caller
would always want to know only the CodeUri of a few functions. Make the
base module more generic: given an asset and a version, return the body.
The primary value the module provides is to give us a consistent default
version.