fix(metricstream): lookup recommended filters
Rather than hardcoding a single filter set into the repo, look it up
by retrieving the SAM manifest. This allows us to roll back (and
forwards) between SAM app releases, and reduces the scope for
misconfiguration. The tradeoff is we are now dependent on an extra HTTP
call to S3 on every plan / apply.